% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_converters.R
\name{as_hcocena}
\alias{as_hcocena}
\title{Convert legacy \code{hcobject} to \code{HCoCenaExperiment}}
\usage{
as_hcocena(hcobject)
}
\arguments{
\item{hcobject}{Legacy list-style analysis object.}
}
\value{
A valid \code{HCoCenaExperiment}.
}
\description{
Convert legacy \code{hcobject} to \code{HCoCenaExperiment}
}
