% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tf.R
\name{check_tf}
\alias{check_tf}
\title{Check Transcription Factor}
\usage{
check_tf(TF)
}
\arguments{
\item{TF}{A string giving the name of the transcription factor to be queried.}
}
\description{
This function leverages the information collected with TF_enrich_network()
to allow the user to query specific transcription factors of interest and see how their top targets are spread across modules.
The goal is to uncover potential co-regulations between clusters.
}
