% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_annotation.R
\name{col_anno_categorical}
\alias{col_anno_categorical}
\title{Categorical Metadata Annotation For The Module Heatmap}
\usage{
col_anno_categorical(variables, variable_label = NULL, type = "abs")
}
\arguments{
\item{variables}{A vector of strings. The length of the vector must equal the number of datasets.
For each dataset, give the name of the column in the corresponding annotation table, that contains the metadata of interest.
E.g., if you want to show sex, and in the first annotationtable the coulm name is "sex" and in the second annotation table it's by mistake labelled as "gender"",
then you set variables = c("sex", "gender").
If the information only exists in some but not all datasets, then set the vector slots of those that don"t have it to NA, e.g., c(NA, "gender").}

\item{variable_label}{A string that describes the meta information used. This string will be used to label the annotation in the heatmap.
If none is provided, the first non-NA string from 'variables' is used as the label.}

\item{type}{A string defining whether meta variables should be converted into percentages ("percent") or absolute ("abs"; default) values}
}
\description{
Annotates the sample groups in the module heatmap with categorical metadata from the annotations.
The annotations will be shown in the heatmap when calling plot_cluster_heatmap() the next time.
}
