% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_annotation.R
\name{col_anno_numerical}
\alias{col_anno_numerical}
\title{Numerical Metadata Annotation For The Module Heatmap}
\usage{
col_anno_numerical(variables, variable_label)
}
\arguments{
\item{variables}{A vector of strings. The length of the vector must equal the number of datasets.
For each dataset, give the name of the column in the corresponding annotation table, that contains the metadata of interest.
E.g., if you want to show age, and in the first annotationtable the coulm name is "age" and in the second annotation table is "Age",
then you set variables = c("age", "Age").
If the information only exists in some but not all datasets, then set the vector slots of those that don"t have it to NA, e.g., c(NA, "Age").}

\item{variable_label}{A string that describes the meta information used. This string will be used to label the annotation in the heatmap.
If none is provided, the first non-NA string from 'variables' is used as the label.}
}
\description{
Annotates the sample groups in the module heatmap with numerical metadata from the annotations.
The annotations will be shown in the heatmap when calling plot_cluster_heatmap() the next time.
}
