% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_single_cluster.R
\name{colour_single_cluster}
\alias{colour_single_cluster}
\title{Colour Single Cluster}
\usage{
colour_single_cluster(cluster)
}
\arguments{
\item{cluster}{A string, giving the colour of the cluster to be highlighted.}
}
\description{
Replots the integrated network, highlighting a given cluster.
}
