% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroup_samples.R
\name{cut_hclust}
\alias{cut_hclust}
\title{Cut Hierarchical Clustering Tree For Sample Regrouping}
\usage{
cut_hclust(
  by = "all",
  set = "all",
  method = "complete",
  k = base::rep(1, base::length(hcobject[["layers"]])),
  save = F
)
}
\arguments{
\item{by}{This parameter defines based on what data to regroup the samples. It can be one of: "all" (the gene expressions of all genes are used (DEFAULT)),
"network" (only the expressions of genes present in the layer's network are used) or "module" (the regrouping is done based on the samples' mean expression patterns across modules).
Please note that in case of by = "network" and by = "module" the output is a heatmap including a dendrogram, while in the case of by = "all" the output is only the dendrogram.
This is due to performance reasons, since heatmaps spanning the entire gene space are often to memory intensive to plot.}

\item{set}{Defines the layers/datasets of which the samples shall be regrouped. Either "all" (provides a hierarchical clustering for all datasets based on which samples can be regrouped (DEFAULT))
or as vector of integers e.g. c(1,3,...), only the layers corresponding to the numbers given in the vector are up for regrouping.}

\item{method}{The method used for the agglomerative clustering, for options refer to the 'hclust()' documentation, DEFAULT is 'complete'.}

\item{k}{A vector of integers giving for each layer that is meant to be regrouped the number of clusters to retrieve after cutting the dendrogram.
Default is a vector of 1s (no cutting, good for first inspection).}

\item{save}{A Boolean, default is FALSE.
Once you have decided on the k-values for cutting the trees and regrouping the samples accordingly, RE-RUN the function, and set this parameter to TRUE.
This will export a new annotation file where the (voi) column now contains the new groups and it will also alter you annotation object in hcobject$data accordingly
(you will find the original groups in a new column labelled (voi)_old ).
The file will be saved in the same directory as the original annotation, bearing the name of the annotation file with the '_regrouped' suffix.
THEN RE-RUN CoCena FROM THE 'run_expression_analysis_2' FUNCTION ONWARD (including that function).
NOTE: New annotation files are always exported without rownames.}
}
\description{
If the variable of interest (voi) does not go well with the clustering of the heatmaps returned by "run_expression_analysis_2" or as seen in the PCA,
new group labels can be assigned to the samples based on the data structure rather than meta information.
This allows you to analyse the data by defining unknown subgroups and thus not solely rely on prior knowledge.
The regrouping can be performed using hierarchical clustering on all genes, the network genes or the modules.
Note: This function can be rerun to try out multiple settings for the k-parameter, if the 'save' parameter is set to FALSE. If it is set to TRUE, the original groups will be overwritten.
Thus, only set 'save' to TRUE once you have decided on your k to cut the tree.
}
