% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_clustering.R
\name{export_clusters}
\alias{export_clusters}
\title{Export Clustering}
\usage{
export_clusters()
}
\description{
Exports your clustering to a text file that can be imported into other hCoCena runs.
Format: Two columns, the first one named gene (containing the gene names) and the second one named color (containing the corresponding cluster color to which the gene belonged).
}
