% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{export_to_cytoscape}
\alias{export_to_cytoscape}
\title{Export Integrated Network To Cytoscape}
\usage{
export_to_cytoscape(name = "my igraph", docker_container = FALSE)
}
\arguments{
\item{name}{A string. The name given to the graph in Cytoscape. Default is "my igraph".}
}
\description{
Attention: Cytoscape Software must be open.
Due to difficulties in the communication between R/RCy3 and Cytoscape, you need to manually stop the function in R as soon as the table of nodes and edges appears in Cytoscape.(by pressing the little stop sign above the console).
}
