% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{export_to_local_folder}
\alias{export_to_local_folder}
\title{Export Integrated Network to local R session
Due to a lacking possibility of communication with Cytoscape from within Docker container, you need to export all necessary information for import into a local R session.}
\usage{
export_to_local_folder()
}
\description{
Export Integrated Network to local R session
Due to a lacking possibility of communication with Cytoscape from within Docker container, you need to export all necessary information for import into a local R session.
}
