% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_to_cluster.R
\name{GeneToCluster}
\alias{GeneToCluster}
\title{Gene To Cluster Dictionary}
\usage{
GeneToCluster(
 
    cluster_information = hcobject[["integrated_output"]][["cluster_calc"]][["cluster_information"]]
)
}
\value{
A data frame with two columns, the first containing gene names as strings, the second containing cluster colours as strings.
}
\description{
The function maps the gene names to their corresponding cluster.
}
