% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_cluster.R
\name{get_cluster_genes}
\alias{get_cluster_genes}
\title{Get Cluster Genes}
\usage{
get_cluster_genes(cluster)
}
\arguments{
\item{cluster}{The colour of the cluster as a string.}
}
\value{
A vector of strings
}
\description{
Returns all genes present in a given cluster as a vector.
}
