% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_scores.R
\name{get_cluster_scores}
\alias{get_cluster_scores}
\title{Get Cluster Scores}
\usage{
get_cluster_scores(save = T)
}
\arguments{
\item{save}{A Boolean. Whether or not to save the plot to PDF (default is TRUE).}
}
\description{
For every gene, the ratio of its edges to genes in the same cluster to its total number of edges is determined.
The corresponding values are returned as a data frame and a box plot is generated showing the scores for each of the clusters.
}
