% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_profiling.R
\name{go_profiling}
\alias{go_profiling}
\title{Gene Ontology Profiling}
\usage{
go_profiling(
  level,
  top,
  ont,
  fast = F,
  padj = "BH",
  clusters = c("all"),
  qval = 0.05
)
}
\arguments{
\item{level}{An integer. From which level in the GO hierarchy the terms should be used for the enrichment.
To see which term resides on which level, please refer to the browse section of AmiGO2 (http://amigo.geneontology.org/amigo/landing).}

\item{top}{Integer. The number of most strongly enriched terms to return per cluster.}

\item{ont}{The subontology to use, three options are available: the molecular functions describing the actions a gene is involved in on a molecular level (“MF”),
cellular component (“CC”), which describes the location within the cell at which a gene exerts its function and biological process (“BP”) i.e., the overall procedure or program a gene is involved in.}

\item{fast}{A Boolean.  If set to TRUE, the enrichment analysis is based on the GO file saved in “supplementary_data” instead of accessing the entire ontology.
In this case, the “level” parameter is ignored. This is much faster at the cost of lower resolution.
If the parameter is set to FALSE (default), the entire database can be searched at any level.
The reference gene list for all enrichments is the union of all genes present in the count files provided for the CoCena analysis.}

\item{padj}{Method to use for multiple testing correction. Can be one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".  Default is "BH" (Benjamini-Hochberg).}

\item{clusters}{Either "all" (default) or a vector of clusters as strings. Defines for which clusters to perform the enrichment.}

\item{qval}{Upper threshold for the adjusted p-value. Default is 0.05.}
}
\description{
Returns a cluster-wise Gene Ontology Database enrichment.
}
