\name{hc_accessors}
\alias{hc_mae}
\alias{hc_config}
\alias{hc_layer_results}
\alias{hc_integration}
\alias{hc_clusters}
\title{Accessors for HCoCenaExperiment}
\description{Extract individual components from an \code{HCoCenaExperiment}.}
\usage{
hc_mae(x)
hc_config(x)
hc_layer_results(x)
hc_integration(x)
hc_clusters(x)
}
\arguments{
\item{x}{An \code{HCoCenaExperiment}.}
}
\value{The requested object component.}
