% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_check_dirs}
\alias{hc_check_dirs}
\title{Run \code{check_dirs()} with S4 state synchronization}
\usage{
hc_check_dirs(hc, create_output_dir = TRUE)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{create_output_dir}{Boolean. If TRUE and \code{dir_output} is missing, create it.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Run \code{check_dirs()} with S4 state synchronization
}
