% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_check_tf}
\alias{hc_check_tf}
\title{Check TF targets (S4 API)}
\usage{
hc_check_tf(hc, TF)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{TF}{Transcription factor symbol.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Check TF targets (S4 API)
}
