% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_colour_single_cluster}
\alias{hc_colour_single_cluster}
\title{Highlight single cluster in network (S4 API)}
\usage{
hc_colour_single_cluster(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{colour_single_cluster()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Highlight single cluster in network (S4 API)
}
