% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{hc_config}
\alias{hc_config}
\alias{hc_config,HCoCenaExperiment-method}
\title{Access configuration from an \code{HCoCenaExperiment}}
\usage{
hc_config(x)

\S4method{hc_config}{HCoCenaExperiment}(x)
}
\arguments{
\item{x}{An \code{HCoCenaExperiment}.}
}
\value{
A \code{HCoCenaConfig} object.
}
\description{
Access configuration from an \code{HCoCenaExperiment}
}
