% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_define_layers}
\alias{hc_define_layers}
\title{Define layers in an \code{HCoCenaExperiment}}
\usage{
hc_define_layers(hc, data_sets = list())
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{data_sets}{Named list of layers, each value a pair (count, annotation).}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Define layers in an \code{HCoCenaExperiment}
}
