% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_export_to_local_folder}
\alias{hc_export_to_local_folder}
\title{Export network to local folder (S4 API)}
\usage{
hc_export_to_local_folder(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{export_to_local_folder()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Export network to local folder (S4 API)
}
