% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_find_hubs}
\alias{hc_find_hubs}
\title{Hub detection (S4 API)}
\usage{
hc_find_hubs(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{find_hubs()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Hub detection (S4 API)
}
