% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_get_module_scores}
\alias{hc_get_module_scores}
\title{Module scores (S4 API)}
\usage{
hc_get_module_scores(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{get_module_scores()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Module scores (S4 API)
}
