% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_highlight_geneset}
\alias{hc_highlight_geneset}
\title{Highlight gene set in network (S4 API)}
\usage{
hc_highlight_geneset(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{highlight_geneset()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Highlight gene set in network (S4 API)
}
