% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_import_layout_from_cytoscape}
\alias{hc_import_layout_from_cytoscape}
\title{Import layout from Cytoscape (S4 API)}
\usage{
hc_import_layout_from_cytoscape(hc)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Import layout from Cytoscape (S4 API)
}
