% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_init}
\alias{hc_init}
\title{Initialize an empty \code{HCoCenaExperiment}}
\usage{
hc_init()
}
\value{
A valid, empty \code{HCoCenaExperiment}.
}
\description{
Initialize an empty \code{HCoCenaExperiment}
}
