% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_init_save_folder}
\alias{hc_init_save_folder}
\title{Run \code{init_save_folder()} with S4 state synchronization}
\usage{
hc_init_save_folder(hc, name, use_output_dir = FALSE)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{name}{Folder name. Use \code{""} to write directly into \code{dir_output}.}

\item{use_output_dir}{Boolean. If TRUE, ignore \code{name} and use \code{dir_output} directly.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Run \code{init_save_folder()} with S4 state synchronization
}
