% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{hc_integration}
\alias{hc_integration}
\alias{hc_integration,HCoCenaExperiment-method}
\title{Access integration payload from an \code{HCoCenaExperiment}}
\usage{
hc_integration(x)

\S4method{hc_integration}{HCoCenaExperiment}(x)
}
\arguments{
\item{x}{An \code{HCoCenaExperiment}.}
}
\value{
A \code{HCoCenaIntegration} object.
}
\description{
Access integration payload from an \code{HCoCenaExperiment}
}
