% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_module_condition_significance}
\alias{hc_module_condition_significance}
\title{Test module differences between conditions (S4 API)}
\usage{
hc_module_condition_significance(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{module_condition_significance()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Test module differences between conditions (S4 API)
}
