% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_pca}
\alias{hc_pca}
\title{PCA plotting (S4 API)}
\usage{
hc_pca(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{PCA()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
PCA plotting (S4 API)
}
