% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_pca_algo_compare}
\alias{hc_pca_algo_compare}
\title{PCA algorithm comparison (S4 API)}
\usage{
hc_pca_algo_compare(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{PCA_algo_compare()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
PCA algorithm comparison (S4 API)
}
