% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_plot_cluster_heatmap}
\alias{hc_plot_cluster_heatmap}
\title{Plot module heatmap (S4 API)}
\usage{
hc_plot_cluster_heatmap(hc, file_name = NULL, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{file_name}{Optional file name passed to \code{plot_cluster_heatmap()}.}

\item{...}{Passed to \code{plot_cluster_heatmap()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Plot module heatmap (S4 API)
}
