% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_plot_cutoffs}
\alias{hc_plot_cutoffs}
\title{Plot cut-off diagnostics (S4 API)}
\usage{
hc_plot_cutoffs(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{plot_cutoffs()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Plot cut-off diagnostics (S4 API)
}
