% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_plot_deg_dist}
\alias{hc_plot_deg_dist}
\title{Plot degree distributions (S4 API)}
\usage{
hc_plot_deg_dist(hc)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Plot degree distributions (S4 API)
}
