% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_plot_gfc_network}
\alias{hc_plot_gfc_network}
\title{Plot network colored by GFC (S4 API)}
\usage{
hc_plot_gfc_network(hc)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Plot network colored by GFC (S4 API)
}
