% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_plot_sample_distributions}
\alias{hc_plot_sample_distributions}
\title{Plot sample distributions (S4 API)}
\usage{
hc_plot_sample_distributions(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{plot_sample_distributions()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Plot sample distributions (S4 API)
}
