% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_read_supplementary}
\alias{hc_read_supplementary}
\title{Load supplementary references (S4 API)}
\usage{
hc_read_supplementary(hc)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Load supplementary references (S4 API)
}
