% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_run_expression_analysis_1}
\alias{hc_run_expression_analysis_1}
\title{Run expression analysis part I (S4 API)}
\usage{
hc_run_expression_analysis_1(
  hc,
  padj = "none",
  export = FALSE,
  import = NULL,
  bayes = FALSE,
  prior = 2,
  alpha = 0.5,
  corr_method = "pearson"
)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{padj}{A String. Defines the method to be used for p-value adjustment. Valid values are "none" (default) or values for "method" in stats::p.adjust.}

\item{export}{A Boolean. If TRUE, correlation values and p-values will be exported.
This can save time if you plan on re-running the analysis since computing pari-wise correlations is a bottleneck of the analysis. Default is FALSE.}

\item{import}{A list. Each slot in the list corresponds to one of the layers (datasets) and is either a vector of two strings (1. path to file holding the correlation matrix and
2. Path to the file holding the p-value matrix) or NA. A list slot is set to NA if for that layer you do not want to import a pre-calculated correlation matrix.
The files do not necessarily have to be exported from a previous run, but can have any kind of origin (created with a different program or method).
For compatibility it is only important, that it is a whitespace-separated text (.txt) file containing a symmetric, numeric matrix.
The first line has to be gene names, there must be no row names, since the first line will be used for column names and row names.
Also, you must provide a matrix with correlation values AND a matrix with corresponding p-values, where cells in the matrices correspond to each other (only a correlation matrix will not be sufficient).
Default is NULL.}

\item{bayes}{Sanchez-Taltavull et al. (2016) suggest superiority of Bayesian correlation analysis to Pearson correlation in some cases.
Therefore, the Pearson correlation values can be weighted with Bayesian correlation values. To do so, set the "bayes"-parameter to TRUE. Default is FALSE, using only Pearson correlations.}

\item{prior}{An integer, either 2 or 3, using prior 2 or 3 for the Bayes weighting as described in "Bayesian correlation analysis for sequence count data" by Sanchez-Taltavull et al. (2016).}

\item{alpha}{A numeric value from 0 to 1. Allows to adjust the strength of the Bayes weighting: For alpha = 0 the Pearson correlation values remain unaltered, for alpha = 1 the Pearson correlation value and the Bayesian correlation value contribute equally to the final correlation.}

\item{corr_method}{Default is "pearson", but can alternatively be set to "spearman" or "rho" in case of single-cell data (according to Skinnider et al., https://www.nature.com/articles/s41592-019-0372-4).}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Run expression analysis part I (S4 API)
}
