% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_set_paths}
\alias{hc_set_paths}
\title{Set input/output paths in an \code{HCoCenaExperiment}}
\usage{
hc_set_paths(
  hc,
  dir_count_data,
  dir_annotation,
  dir_reference_files,
  dir_output
)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{dir_count_data}{Path to count data.}

\item{dir_annotation}{Path to annotation data.}

\item{dir_reference_files}{Path to reference files.}

\item{dir_output}{Output directory.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Set input/output paths in an \code{HCoCenaExperiment}
}
