% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_suggest_topvar}
\alias{hc_suggest_topvar}
\title{Suggest top variable genes (S4 API)}
\usage{
hc_suggest_topvar(hc)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Suggest top variable genes (S4 API)
}
