% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_tf_overrep_module}
\alias{hc_tf_overrep_module}
\title{TF enrichment per module (S4 API)}
\usage{
hc_tf_overrep_module(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{TF_overrep_module()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
TF enrichment per module (S4 API)
}
