% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_tf_overrep_network}
\alias{hc_tf_overrep_network}
\title{TF enrichment network-wide (S4 API)}
\usage{
hc_tf_overrep_network(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{TF_overrep_network()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
TF enrichment network-wide (S4 API)
}
