% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_update_clustering_algorithm}
\alias{hc_update_clustering_algorithm}
\title{Update clustering algorithm (S4 API)}
\usage{
hc_update_clustering_algorithm(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{update_clustering_algorithm()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Update clustering algorithm (S4 API)
}
