% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_visualize_gene_expression}
\alias{hc_visualize_gene_expression}
\title{Visualize gene expression (S4 API)}
\usage{
hc_visualize_gene_expression(hc, ...)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}

\item{...}{Passed to \code{visualize_gene_expression()}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Visualize gene expression (S4 API)
}
