% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_api.R
\name{hc_write_session_info}
\alias{hc_write_session_info}
\title{Write session info (S4 API)}
\usage{
hc_write_session_info(hc)
}
\arguments{
\item{hc}{A \code{HCoCenaExperiment}.}
}
\value{
Updated \code{HCoCenaExperiment}.
}
\description{
Write session info (S4 API)
}
