% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HCoCenaClasses}
\alias{HCoCenaClasses}
\alias{HCoCenaConfig-class}
\alias{HCoCenaReferences-class}
\alias{HCoCenaLayerResult-class}
\alias{HCoCenaIntegration-class}
\alias{HCoCenaExperiment-class}
\title{hCoCena S4 Classes}
\description{
S4 container classes used by the object-oriented hCoCena API.
}
\keyword{classes}
