% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_geneset.R
\name{highlight_geneset}
\alias{highlight_geneset}
\title{Highlight Geneset}
\usage{
highlight_geneset(gene_set, col = "black")
}
\arguments{
\item{gene_set}{A vector of gene symbols (as strings) to be highlighted in the network.}

\item{col}{A string specifying the color with which the nodes of the genes are framed to highlight them. Default is "black".}
}
\description{
Receives and highlights a list of genes in the integrated network.
}
