% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{import_layout_from_cytoscape}
\alias{import_layout_from_cytoscape}
\title{Import Layout From Cytoscape}
\usage{
import_layout_from_cytoscape()
}
\description{
Imports the layout of a network currently open in Cytoscape.
}
