% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{import_layout_from_local_folder}
\alias{import_layout_from_local_folder}
\title{Import network layout from file to your Docker container
Imports the layout generated within a local R session and Cytoscape back into the Docker container}
\usage{
import_layout_from_local_folder()
}
\description{
Import network layout from file to your Docker container
Imports the layout generated within a local R session and Cytoscape back into the Docker container
}
