% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_object.R
\name{init_object}
\alias{init_object}
\title{Initialise hCoCena Object}
\usage{
init_object(envo = .GlobalEnv)
}
\description{
Creates an object called 'hcobject' in the global envirentment that will be used by hCoCena throughout the analysis.
}
