% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_save_folder.R
\name{init_save_folder}
\alias{init_save_folder}
\title{Creates a Save Folder}
\usage{
init_save_folder(name)
}
\arguments{
\item{name}{The name of the folder to be created.
Use \code{""} to write outputs directly into \code{dir_output} without creating a subfolder.}
}
\description{
A folder with the given name is created in the output directory. All analysis outputs will be saved to this folder.
}
