% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_wd.R
\name{init_wd}
\alias{init_wd}
\title{Initialise the Working Directory}
\usage{
init_wd(dir_count_data, dir_annotation, dir_reference_files, dir_output)
}
\arguments{
\item{dir_count_data}{A string specifying the path of the folder containing the count data files. All count files should be found in the same directory. For further specifications regarding the count files, see the documentation of read_data().}

\item{dir_annotation}{A string specifying the path of the folder containing the annotation data files. All annotation files should be found in the same directory. For further specifications regarding the annotation files, see the documentation of read_data().}

\item{dir_reference_files}{A string specifying the path of the folder containing the reference files needed for some analysis steps. The files can be found on the GitHub repository under 'reference files'. For details on the files and how to exchange them, refer to the README in the repository.}

\item{dir_output}{A string specifying the path of the folder in which hCoCena will create a new save folder for the analysis outputs.}
}
\description{
init_wd sets up from which directory to read the count data, the annotation data, the reference files and it defines in which directory to create a save folder that all outputs will be written to.
}
\seealso{
\code{\link[=read_data]{read_data()}}
}
