% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_correlation.R
\name{meta_correlation_num}
\alias{meta_correlation_num}
\title{Numeric Metadata Correaltion}
\usage{
meta_correlation_num(set, meta, p_val = 0.05)
}
\arguments{
\item{set}{An integer. The number of the datset for which to perform the calculation.}

\item{meta}{A vector of strings. The names of the numeric annotation column(s) which to correlate to the cluster expression patterns.}

\item{p_val}{The maximum p-value to determine a correlation as significant. Default is 0.05. Non-significant correlations are shown in grey.}
}
\description{
Calculates a matrix where columns are names of the meta categories and rows are clusters. Cells contain the Pearson correlation value between
a) the mean expressions of cluster genes in each sample with
b) the numeric meta value in each sample.
}
