% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_comparison.R
\name{network_comparison_2}
\alias{network_comparison_2}
\title{Compare degree distribution of gene set}
\usage{
network_comparison_2(net1, net2, as = "igraph", gene_vec)
}
\arguments{
\item{net1}{An igraph object of the first network. hCoCena provides its integrated network as igraph objects. You can find it here: hcobject$integrated_output$merged_net.
If you want to create a network manually with external data, please refer to the igraph documentation: https://igraph.org/r/.}

\item{net2}{See 'net1', only for network 2.}

\item{as}{In the future we plan to provide either an igraph object or an edge list. For now, only the igraph option is available.}

\item{gene_vec}{A vector of gene symbols you wish to investigate. Gene symbols must be provided as strings, e.g., c( "YME1L1", "SLC2A5", "SIAH2", "GPI", "IL10RB").}
}
\value{
The ggplot object for replotting/re-sizing/modification etc. and the data used to plot the ggplot. Output can be found under hcobject$satellite_outputs$network_comparison_2
}
\description{
This function accepts two networks and a set of genes. It then calculates the degree-distribution of each gene in both networks and also the Jaccard-Index
of each gene's neighbourhoods in the two networks. The results are visualized in a 2D dot plot.
This allows the comparison of the two networks with respect to the connectivity of selected genes.
}
