% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_plot.R
\name{plot_cutoffs}
\alias{plot_cutoffs}
\title{Plot of Cutoff Statistics}
\usage{
plot_cutoffs(
  interactive = T,
  hline = list(R.squared = NULL, no_edges = NULL, no_nodes = NULL, no_networks = NULL)
)
}
\arguments{
\item{interactive}{Boolean. If TRUE (default) plot is created using plotly including and interactive cutoff slider. If FALSE, plot is created as a static ggplot.}

\item{hline}{A list with four slots ("R.squared", "no_edges", "no_nodes", "no_networks") each of which can be set either to NULL (default) or a number to introduce a horizontal line for orientation at that value in the respective plot. Only used in the non-interactive plot.}
}
\description{
Plots the R-squared value, number of edges, number of genes and number of networks for different cut-offs.
}
