% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deg_dist.R
\name{plot_deg_dist}
\alias{plot_deg_dist}
\title{Plot Degree Distribution}
\usage{
plot_deg_dist()
}
\description{
A function that plot the degree distribution of each network after being cut with the set cutoff values.
}
