% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_integrated_network.R
\name{plot_integrated_network}
\alias{plot_integrated_network}
\title{Plot Integrated Network}
\usage{
plot_integrated_network(
  layout = NULL,
  layout_algorithm = NULL,
  gene_labels = NULL,
  save = T,
  label_offset = 50
)
}
\arguments{
\item{layout}{NULL or a precalculated layout. If NULL (default) the layout is calculated.}

\item{layout_algorithm}{Optional layout algorithm for this plot only.
If \code{NULL} (default), the algorithm from \code{set_global_settings()} is used.
Supported values are \code{"layout_with_fr"}, \code{"layout_with_stress"},
\code{"layout_with_sparse_stress"}, \code{"layout_with_drl"}, \code{"layout_with_kk"} and \code{"cytoscape"}.}

\item{gene_labels}{A vector of character strings containing the names of genes to be labelled. Labels will be placed with an offset, which can be adjusted using the label_offset parameter.}

\item{save}{A Boolean stating whether of not to save the network to a PDF. Default is TRUE.}

\item{label_offset}{The offset of provided gene labels to the network center. Default is 50, depends on network size.}
}
\description{
Creates an igraph of the integrated network.
}
