% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_factor_enrichment.R
\name{plot_TF_enrichment}
\alias{plot_TF_enrichment}
\title{Visualize Transcription Factor Enrichment Results}
\usage{
plot_TF_enrichment()
}
\description{
Results are visualized as circular plots, one for each cluster. Each cell in the plot is labelled with a gene name.
The names of the transcription factors are highlighted in turquoise, those of target genes are written in black.
The colour of the cell symbolises in which module the target or the transcription factor can be found in.
If a gene is a target of one the enriched transcription factors, a link is connecting their cells.
If there is an edge in the constructed co-expression network connecting those two genes, that link is non-transparent.
If the TF/target-pair was found by ChEA3 but is not represented by an edge in the network, the link is depicted as slightly transparent.
}
