% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome_profiling.R
\name{reactome_profiling}
\alias{reactome_profiling}
\title{Reactome Profiling}
\usage{
reactome_profiling(top, padj = "BH", clusters = c("all"), qval = 0.05)
}
\arguments{
\item{top}{Integer. The number of most strongly enriched terms to return per cluster.}

\item{padj}{Method to use for multiple testing correction. Can be one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".  Default is "BH" (Benjamini-Hochberg).}

\item{clusters}{Either "all" (default) or a vector of clusters as strings. Defines for which clusters to perform the enrichment.}

\item{qval}{Upper threshold for the adjusted p-value. Default is 0.05.}
}
\description{
Returns a cluster-wise Reactome Database enrichment.
}
