% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Function To Read All Count And Annotation Files}
\usage{
read_data(
  sep_counts = "\\t",
  sep_anno = "\\t",
  gene_symbol_col = NULL,
  sample_col = NULL,
  count_has_rn = TRUE,
  anno_has_rn = TRUE
)
}
\arguments{
\item{sep_counts}{The separator of the count files. Default is tab separated files. Ignore when loading data from objects instead of files.}

\item{sep_anno}{The separator of the annotation files. Default is tab separated files. Ignore when loading data from objects instead of files.}

\item{gene_symbol_col}{A String. Name of the column that contains the gene symbols. Ignore when loading data from objects instead of files.}

\item{sample_col}{A String. Name of the column that contains the sample IDs. Ignore when loading data from objects instead of files.}

\item{count_has_rn}{A Boolean. Whether or not the count file has rownames. Default is TRUE. Ignore when loading data from objects instead of files.}

\item{anno_has_rn}{A Boolean. Whether or not the annotation file has rownames. Default is TRUE. Ignore when loading data from objects instead of files.}
}
\description{
The function loads the count and annotation data for each layer and saves it in the hCoCena-Object's "data" slot.
}
