% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_supplementary.R
\name{read_supplementary}
\alias{read_supplementary}
\title{Read Supplementary Data}
\usage{
read_supplementary()
}
\description{
Function to read and collect the supplementary files set in set_supp_files().
}
