% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_topvar.R
\name{suggest_topvar}
\alias{suggest_topvar}
\title{Suggest Top Most Variant Genes}
\usage{
suggest_topvar()
}
\description{
For each dataset, calculate the inflection points of the logged variance of gene expression values.
Returned values may be used for "topvar" in set_layer_settings().
}
