% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_factor_enrichment.R
\name{TF_enrich_all}
\alias{TF_enrich_all}
\title{Network-Wide Transcription Factor Overrepresentation Analysis}
\usage{
TF_enrich_all(topTF = 100, topTarget = 30)
}
\arguments{
\item{topTF}{The number of transcription factors with the highest number of enricht targets in the network. Default is 100.}

\item{topTarget}{Per transcription factor the number of top most enriched targets to return. Default is 30.}
}
\description{
Returns the transcription factors that have the most enriched targets network-wide, including their targets.
}
