% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_factor_enrichment.R
\name{TF_overrep}
\alias{TF_overrep}
\title{Transfiction Factor Overrepresentation Analysis}
\usage{
TF_overrep()
}
\description{
Runs a ChEA3 transcription factor enrichment analysis for every gene module detected in the network.
It filters the ranked enriched transcription factors for the 5 highest ranking ones and their 5 highest ranking targets.
MeanRank was chosen as a ranking method.
}
