% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription_factor_enrichment.R
\name{TF_overrep_module}
\alias{TF_overrep_module}
\title{Transcription Factor Over-representation Analysis}
\usage{
TF_overrep_module(clusters = "all", topTF = 5, topTarget = 5)
}
\arguments{
\item{clusters}{Either "all" (default) or a vector of clusters as strings. Defines for which clusters to perform the analysis.}

\item{topTF}{Integer. The number of top ranking TFs to return per cluster. Default is 5.}

\item{topTarget}{Integer. The number of top ranking targets to return per TF. Default is 5.}
}
\description{
Runs a ChEA3 transcription factor (TF) enrichment analysis for every selected gene module.
It filters the ranked enriched TFs for the top highest ranking ones and their top highest ranking targets.
MeanRank was chosen as a ranking method.
Results are visualized as circular plots, one for each module.
The outer circle highlights TFs (grey) and their targets (white). Within the inner circle, the colour of the cell is colored according to the module the respective gene can be found in.
If a gene is a target of one of the enriched transcription factors, a dashed link is connecting their cells.
If there is an edge in the constructed co-expression network connecting those two genes, that link is solid.
}
