% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_clustering_algorithm.R
\name{update_clustering_algorithm}
\alias{update_clustering_algorithm}
\title{Change Or Update The Clustering Algorithm}
\usage{
update_clustering_algorithm(new_algo = NULL, gtc = NULL)
}
\arguments{
\item{new_algo}{A string (name of the new algorithm), or NULL (if gtc is provided)}

\item{gtc}{A user defined clsutering can be provided as long as it has the same format as the output of GeneToCluster(): A data frame with two columns, the first containing gene names as strings, the second containing cluster colours as strings.}
}
\description{
The clustering can be updated to another algortihm (one of "cluster_louvain", "cluster_fast_greedy", "cluster_infomap", "cluster_walktrap", "cluster_label_prop", "cluster_leiden") or to a user defined clsutering using the 'gtc' parameter.
}
