% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_gene_expression.R
\name{visualize_gene_expression}
\alias{visualize_gene_expression}
\title{Visualize Gene Expression}
\usage{
visualize_gene_expression(genes, name = NULL, width = 15, height = 10)
}
\arguments{
\item{genes}{A vector of strings giving the genes to be plotted.}

\item{name}{A string giving the name for the save file WITHOUT file ending.}

\item{width}{Width of the PDF, default ist 15, only change if plots overlap in PDF.}

\item{height}{Height of the PDF, default ist 10, only change if plots overlap in PDF.}
}
\description{
Plots the mean expression values per condition for the given genes for each dataset as a heatmap. Values are scaled across rows.
}
