% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_cluster.R
\name{zoom2}
\alias{zoom2}
\title{Zoom Into CLuster Part II}
\usage{
zoom2(cutoff_vec, cluster_algo = "cluster_leiden")
}
\arguments{
\item{cutoff_vec}{A vector of cutoff values. The vector's length must be equal to the number of datasets. The order in which the cutoffs are set must correspond to the order in which the layers were declared in define_layers().}

\item{cluster_algo}{The clustering algorithm to be used. The choise is between "cluster_leiden" (default), "cluster_louvain", "cluster_label_prop", "cluster_fast_greedy",
"cluster_infomap", "cluster_walktrap" and "auto" (in which case all are tested and the one with the highest modularity is chosen).}
}
\description{
Filters Cluster genes for set cutoff, integrates layers (if necessary), re-clusters and re-plots cluster-network
}
