% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_cluster.R
\name{zoom_settings}
\alias{zoom_settings}
\title{Cluster Zoom Settings

Sets the parameters to zoom into a specific cluster.}
\usage{
zoom_settings(
  cluster,
  min_nodes = 15,
  mode = "u",
  multi_edges = "min",
  no_of_iterations = 10,
  max_cluster_count_per_gene = 1,
  min_nodes_number_for_cluster = 5
)
}
\arguments{
\item{cluster}{The colour of the cluster as a string.}

\item{min_nodes}{An integer. The minimum number of nodes to comprise a cluster, default is 15.}

\item{mode}{How to integrate the networks of the different layers (can be ignored, if there is only one dataset). Default is 'u' (integration by union), alternatively you can set 'i' (integration by intersection).}

\item{multi_edges}{One of “min” (default), “mean” or “max” resulting in the simplification of the multigraph by using the minimum, the mean or the maximum edge weight among the multiple edges, respectively.
Multiple edges occur when the edge was present in more than one datset.}

\item{no_of_iterations}{How often to repeat the clsutering algorithm. Default is 10.}

\item{max_cluster_count_per_gene}{The maximum number of clsuters a gene can be associated with over the different iteration before it is discared into cluster 'white'. Default is 1.}

\item{min_nodes_number_for_cluster}{The minimum number of nodes to comprise a cluster in the network. Default is 5.}
}
\description{
Cluster Zoom Settings

Sets the parameters to zoom into a specific cluster.
}
