% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroup_cluster_heatmap.R
\name{change_grouping_parameter}
\alias{change_grouping_parameter}
\title{Change Grouping Parameter}
\usage{
change_grouping_parameter(group_by)
}
\arguments{
\item{group_by}{A string giving the new grouping variable. This must be a name of a column present in all annotation files.}
}
\description{
The variable by which the samples are grouped and based on which the GFCs are calculated can be changed and the cluster heatmap will be replotted.
This is particularly useful in cases where different variables are potential candidates for driving the genes’ expression changes in the data
and an explorative approach is required to decide on the most suitable one.
Note that any previously generated column annotation will not be plotted, since the grouping will change.
If you eventually decide on another grouping variable, please run the analysis again entirely with the changed "voi" from the very beginning.
}
