% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dirs.R
\name{check_dirs}
\alias{check_dirs}
\title{Fixes Directories}
\usage{
check_dirs()
}
\description{
Iteratively calls fix_dir() on the provided working directory paths to fix them if necessary or throw an error if they are invalid.
}
