% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{export_to_cytoscape}
\alias{export_to_cytoscape}
\title{Export Integrated Network To Cytoscape}
\usage{
export_to_cytoscape(name = "my igraph")
}
\arguments{
\item{name}{A string. The name given to the graph in Cytoscape. Default is "my igraph".}
}
\description{
Attention: Cytoscape Software must be open.
Due to poor communication between R and Cytoscape, the exports are usually quite slow. To fasten things up, after running the function, wait until the "create view" button appears in Cytoscape.
Press it and wait for the network to render. Then INTERRUPT the function in R (by pressing the little stop sign above the console).
}
