% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{functional_enrichment}
\alias{functional_enrichment}
\title{Functional enrichment analysis}
\usage{
functional_enrichment(
  gene_sets = c("Go", "Kegg", "Hallmark", "Reactome"),
  top = 5,
  clusters = c("all"),
  padj = "BH",
  qval = 0.05
)
}
\arguments{
\item{gene_sets}{A vector. The names of databases enrichment should be performed for. Choose one or multiple of "Go", "Kegg", "Hallmark", and/or "Reactome".
Available databases depend on supplement files previously set}

\item{top}{Integer. The number of most strongly enriched terms to return per cluster. Default is 5.}

\item{clusters}{Either "all" (default) or a vector of clusters as strings. Defines for which clusters to perform the enrichment.}

\item{padj}{Method to use for multiple testing correction. Can be one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".  Default is "BH" (Benjamini-Hochberg).}

\item{qval}{Upper threshold for the adjusted p-value. Default is 0.05.}
}
\description{
Returns a cluster-wise Gene Ontology Database enrichment.
}
