% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{import_layout_from_local_folder}
\alias{import_layout_from_local_folder}
\title{Import network layout from file to your Docker container
Imports the layout generated within a local R session and Cytoscape back into the Docker container}
\usage{
import_layout_from_local_folder(
  file = base::paste0(hcobject[["working_directory"]][["dir_output"]],
    hcobject[["global_settings"]][["save_folder"]], "/network_layout.csv")
)
}
\arguments{
\item{file}{Exact path containing the Cytoscape layout. Default path is set to the save folder defined in the global settings.}
}
\description{
Import network layout from file to your Docker container
Imports the layout generated within a local R session and Cytoscape back into the Docker container
}
