% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge Clusters}
\usage{
merge_clusters(k = 1, save = F, method = "complete")
}
\arguments{
\item{k}{An integer. The number of desired clusters created by cutting the clustering tree.}

\item{save}{Boolean. If FALSE (default) cutting is showcased but not saved. Set to TRUE once you have found the right clsutering and overwrite the old cluster (cannot be reversed).}

\item{method}{Method used for clustering. Default is "complete".}
}
\description{
Merge similar clusters in module heatmap.
}
