% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_heatmap.R
\name{plot_cluster_heatmap}
\alias{plot_cluster_heatmap}
\title{Plot Cluster Heatmap}
\usage{
plot_cluster_heatmap(
  col_order = NULL,
  row_order = NULL,
  cluster_columns = T,
  cluster_rows = T,
  k = 0,
  return_HM = T,
  cat_as_bp = NULL,
  file_name = "module_heatmap.pdf"
)
}
\arguments{
\item{col_order}{Defines the order in which the sample groups (conditions) appear in the heatmap.
Accepts a vector of strings giving the conditions in their desired order, default is NULL (automatic order, determined by clustering).
If the parameter cluster_columns is not set to FALSE, this order will be overwritten.}

\item{row_order}{Like col_order but with cluster names.}

\item{cluster_columns}{A Boolean, whether or not to cluster the columns of the heatmap. Default is TRUE, overwrites col_order.}

\item{cluster_rows}{Like cluster_columns but for rows.}

\item{k}{The resulting cluster_columns tree is cut into k groups. Default is 0 (no cutting).}

\item{return_HM}{A Boolean whether of not to return the ComplexHeatmap object to hcobject$integrated_output$cluster_calc$heatmap_cluster in addition to plotting it. Default is TRUE.}

\item{cat_as_bp}{A vector of Booleans which length is equivalent to the number of categorical meta data annotations created with the satellite functions.
Each Boolean states whether or not the categorical variable should be annotated as a bar plot (TRUE) or as a line plot (FALSE).
If you did not perform any meta data annotation, ignore this parameter.}

\item{file_name}{A string giving the name of the file (with .pdf ending) to which the heatmap should be written. Default is "module_heatmap.pdf".}
}
\description{
Plots a heatmap with sample groups as columns and gene clusters as rows. The cells are coloured according to the mean GFC of a given cluster in the respective sample group.
If categorical or numerical metadata annotations or user-defined enrichments have been created with satellite functions, they will be incorporated into the heatmap as column/row annotations.
For the available options check out the satellite functions.
}
