% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_supp_files.R
\name{set_supp_files}
\alias{set_supp_files}
\title{Set Supplementary Files}
\usage{
set_supp_files(
  Tf = NULL,
  Hallmark = NULL,
  Go = NULL,
  Kegg = NULL,
  Reactome = NULL,
  ...
)
}
\arguments{
\item{Tf}{A file name. The file must contain a list of at least two columns. One column should be titled with the organism from which the datasets originate and contain gene names, the other column must always be the last column and contain the information if the gene is a transcription factor ("TF"),
a co-factor ("Co_factor"), a chromatin remodelling protein ("Chromatin_remodeller") or a ribonucleic acid binding protein ("RNBP"). The name of the last column may vary. An exmemplary file for mouse and human is found in the Reference Files Folder in the repository.}

\item{Go}{A file name for a .gmt gene ontology file. You can find one in the reference file folder in the repository.}

\item{Kegg}{A file name for a .gmt KEGG database file. You can find one in the reference file folder in the repository.}

\item{Reactome}{A file name for a .gmt Reactomte database file. You can find one in the reference file folder in the repository.}

\item{Hall}{A file name for a .gmt MSigDB hallmark gene set file. You can find one in the reference file folder in the repository.}
}
\description{
Receives the file names of the supplementary files
}
\details{
Custom database files can be used in a .csv or .gmt file format. .csv files have the following format:
\enumerate{
\item column (named 'term'): name of the pathway
\item column (named 'gene'): gene symbol of the respective pathway defined in the first column
Note: Only one gene symbol per row is allowed, thus, in cases where multiple genes are associated with a pathway generate one row per gene symbol!
}
}
